Move multi threaded builder struct
