+++ src/binary_tree/builder.rs
@@ -4,6 +4,7 @@ use std::fmt::Debug;
 use super::{BinaryTree, Coordinate, Mergeable, Node, MIN_HEIGHT};
 
 mod multi_threaded;
+use multi_threaded::MultiThreadedBuilder;
 
 mod single_threaded;
 use single_threaded::SingleThreadedBuilder;
@@ -25,15 +26,7 @@ pub struct InputLeafNode<C> {
 }
 
 // -------------------------------------------------------------------------------------------------
-// Tree builder.
-
-pub struct MultiThreadedBuilder<C>
-where
-    C: Clone,
-{
-    height: u8,
-    leaf_nodes: Vec<Node<C>>,
-}
+// Implementation.
 
 impl<C> TreeBuilder<C>
 where
@@ -75,102 +68,6 @@ where
     }
 }
 
-impl<C> MultiThreadedBuilder<C>
-where
-    C: Clone + Mergeable,
-{
-    fn new(parent_builder: TreeBuilder<C>) -> Result<Self, TreeBuildError> {
-        use super::num_bottom_layer_nodes;
-
-        // require certain fields to be set
-        let input_leaf_nodes = parent_builder
-            .leaf_nodes
-            .ok_or(TreeBuildError::NoLeafNodesProvided)?;
-        let height = parent_builder
-            .height
-            .ok_or(TreeBuildError::NoHeightProvided)?;
-
-        let max_leaf_nodes = num_bottom_layer_nodes(height);
-        if input_leaf_nodes.len() as u64 > max_leaf_nodes {
-            return Err(TreeBuildError::TooManyLeaves);
-        }
-
-        // TODO need to parallelize this, it's currently the same as the single-threaded
-        // version Construct a sorted vector of leaf nodes and perform parameter
-        // correctness checks.
-        let mut leaf_nodes = {
-            // Translate InputLeafNode to Node.
-            let mut leaf_nodes: Vec<Node<C>> = input_leaf_nodes
-                .into_iter()
-                .map(|leaf| leaf.to_node())
-                .collect();
-
-            // Sort by x_coord ascending.
-            leaf_nodes.sort_by(|a, b| a.coord.x.cmp(&b.coord.x));
-
-            // Make sure all x_coord < max.
-            if leaf_nodes
-                .last()
-                .is_some_and(|node| node.coord.x >= max_leaf_nodes)
-            {
-                return Err(TreeBuildError::InvalidXCoord);
-            }
-
-            // Ensure no duplicates.
-            let duplicate_found = leaf_nodes
-                .iter()
-                .fold(
-                    (max_leaf_nodes, false),
-                    |(prev_x_coord, duplicate_found), node| {
-                        if duplicate_found || node.coord.x == prev_x_coord {
-                            (0, true)
-                        } else {
-                            (node.coord.x, false)
-                        }
-                    },
-                )
-                .1;
-
-            if duplicate_found {
-                return Err(TreeBuildError::DuplicateLeaves);
-            }
-
-            leaf_nodes
-        };
-
-        Ok(MultiThreadedBuilder { height, leaf_nodes })
-    }
-
-    pub fn build<F>(self, padding_node_generator: F) -> Result<BinaryTree<C>, TreeBuildError>
-    where
-        C: Debug + Send + 'static,
-        F: Fn(&Coordinate) -> C + Send + 'static + Sync,
-    {
-        use std::sync::Arc;
-
-        let height = self.height;
-        let x_coord_min = 0;
-        let x_coord_max = 2u64.pow(height as u32 - 1) - 1;
-        let y = height - 1;
-
-        let root = multi_threaded::build_node(
-            x_coord_min,
-            x_coord_max,
-            y,
-            height,
-            self.leaf_nodes,
-            Arc::new(padding_node_generator),
-        );
-
-        let store = HashMap::new();
-
-        Ok(BinaryTree {
-            root,
-            store,
-            height,
-        })
-    }
-}
 
 use thiserror::Error;
