+++ src/binary_tree/builder/multi_threaded.rs
@@ -1,4 +1,5 @@
 use std::fmt::Debug;
+use std::collections::HashMap;
 
 use std::sync::Arc;
 use std::thread;
@@ -7,6 +8,114 @@ use super::super::{
     num_bottom_layer_nodes, Coordinate, LeftSibling, MatchedPair, Mergeable, Node, NodeOrientation,
     RightSibling, Sibling,
 };
+use super::{TreeBuilder, TreeBuildError, BinaryTree};
+
+// -------------------------------------------------------------------------------------------------
+// Main struct.
+
+pub struct MultiThreadedBuilder<C>
+where
+    C: Clone,
+{
+    height: u8,
+    leaf_nodes: Vec<Node<C>>,
+}
+
+impl<C> MultiThreadedBuilder<C>
+where
+    C: Clone + Mergeable,
+{
+    pub fn new(parent_builder: TreeBuilder<C>) -> Result<Self, TreeBuildError> {
+        // require certain fields to be set
+        let input_leaf_nodes = parent_builder
+            .leaf_nodes
+            .ok_or(TreeBuildError::NoLeafNodesProvided)?;
+        let height = parent_builder
+            .height
+            .ok_or(TreeBuildError::NoHeightProvided)?;
+
+        let max_leaf_nodes = num_bottom_layer_nodes(height);
+        if input_leaf_nodes.len() as u64 > max_leaf_nodes {
+            return Err(TreeBuildError::TooManyLeaves);
+        }
+
+        // TODO need to parallelize this, it's currently the same as the single-threaded
+        // version Construct a sorted vector of leaf nodes and perform parameter
+        // correctness checks.
+        let mut leaf_nodes = {
+            // Translate InputLeafNode to Node.
+            let mut leaf_nodes: Vec<Node<C>> = input_leaf_nodes
+                .into_iter()
+                .map(|leaf| leaf.to_node())
+                .collect();
+
+            // Sort by x_coord ascending.
+            leaf_nodes.sort_by(|a, b| a.coord.x.cmp(&b.coord.x));
+
+            // Make sure all x_coord < max.
+            if leaf_nodes
+                .last()
+                .is_some_and(|node| node.coord.x >= max_leaf_nodes)
+            {
+                return Err(TreeBuildError::InvalidXCoord);
+            }
+
+            // Ensure no duplicates.
+            let duplicate_found = leaf_nodes
+                .iter()
+                .fold(
+                    (max_leaf_nodes, false),
+                    |(prev_x_coord, duplicate_found), node| {
+                        if duplicate_found || node.coord.x == prev_x_coord {
+                            (0, true)
+                        } else {
+                            (node.coord.x, false)
+                        }
+                    },
+                )
+                .1;
+
+            if duplicate_found {
+                return Err(TreeBuildError::DuplicateLeaves);
+            }
+
+            leaf_nodes
+        };
+
+        Ok(MultiThreadedBuilder { height, leaf_nodes })
+    }
+
+    pub fn build<F>(self, padding_node_generator: F) -> Result<BinaryTree<C>, TreeBuildError>
+    where
+        C: Debug + Send + 'static,
+        F: Fn(&Coordinate) -> C + Send + 'static + Sync,
+    {
+        let height = self.height;
+        let x_coord_min = 0;
+        let x_coord_max = 2u64.pow(height as u32 - 1) - 1;
+        let y = height - 1;
+
+        let root = build_node(
+            x_coord_min,
+            x_coord_max,
+            y,
+            height,
+            self.leaf_nodes,
+            Arc::new(padding_node_generator),
+        );
+
+        let store = HashMap::new();
+
+        Ok(BinaryTree {
+            root,
+            store,
+            height,
+        })
+    }
+}
+
+// -------------------------------------------------------------------------------------------------
+// Supporting functions, structs, etc.
 
 /// Returns the index `i` in `nodes` where `nodes[i].coord.x <= x_coord_mid`
 /// but `nodes[i+1].coord.x > x_coord_mid`.
@@ -90,6 +199,10 @@ impl<C: Clone> RightSibling<C> {
     }
 }
 
+// -------------------------------------------------------------------------------------------------
+// Build algorithm.
+
+// TODO need more docs
 /// Recursive multi-threaded function for building a node by exploring the tree
 /// from top-to-bottom.
 ///
